Adapt to VS 2015

+++ FarNet/FarNetMan/DialogControls.cpp
@@ -126,7 +126,7 @@ int FarControl::GetSelected()
 void FarControl::SetSelected(int value)
 {
 	if (_dialog->_hDlg != INVALID_HANDLE_VALUE)
-		Info.SendDlgMessage(_dialog->_hDlg, DM_SETCHECK, Id, (void*)value);
+		Info.SendDlgMessage(_dialog->_hDlg, DM_SETCHECK, Id, (void*)(__int64)value);
 	else
 		_selected = value;
 }
